requirements.txt
folium==0.14.0
pandas==1.5.3
streamlit==1.17.0
pillow==22.2.1
inflection==0.5.1
streamlit_folium==0.11.0
plotly==5.13.0